refactor: extract physics functions into physics.py
Move all coefficient computation functions (line-source, point-source,
reciprocity, CSD) and SegmentGeometry into bluerecording/physics.py.
weights.py keeps orchestration, H5 I/O, and electrode parsing.

Backward-compatible aliases are re-exported from weights.py so
existing test imports continue to work.
